Move all profiles into core
