+++ adjuster/src/test/java/org/jboss/pnc/reqour/adjust/provider/GradleProviderTest.java
@@ -10,7 +10,7 @@ import jakarta.inject.Inject;
 import org.assertj.core.data.MapEntry;
 import org.jboss.pnc.reqour.adjust.AdjustTestUtils;
 import org.jboss.pnc.reqour.adjust.config.ReqourAdjusterConfig;
-import org.jboss.pnc.reqour.adjust.profiles.GradleAdjustProfile;
+import org.jboss.pnc.reqour.common.profile.GradleAdjustProfile;
 import org.jboss.pnc.reqour.common.utils.IOUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
