+++ adjuster/src/test/java/org/jboss/pnc/reqour/adjust/provider/MvnProviderTest.java
@@ -10,7 +10,7 @@ import jakarta.inject.Inject;
 import org.assertj.core.data.MapEntry;
 import org.jboss.pnc.reqour.adjust.AdjustTestUtils;
 import org.jboss.pnc.reqour.adjust.config.ReqourAdjusterConfig;
-import org.jboss.pnc.reqour.adjust.profiles.MvnAdjustProfile;
+import org.jboss.pnc.reqour.common.profile.MvnAdjustProfile;
 import org.jboss.pnc.reqour.common.utils.IOUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
