+++ adjuster/src/test/java/org/jboss/pnc/reqour/adjust/service/AdjustmentPusherTest.java
@@ -9,7 +9,7 @@ import io.quarkus.test.junit.TestProfile;
 import jakarta.inject.Inject;
 import org.apache.commons.io.FileUtils;
 import org.eclipse.jgit.api.errors.GitAPIException;
-import org.jboss.pnc.reqour.adjust.profiles.AdjustProfile;
+import org.jboss.pnc.reqour.common.profile.AdjustProfile;
 import org.jboss.pnc.reqour.common.CloneTestUtils;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
