+++ adjuster/src/test/java/org/jboss/pnc/reqour/adjust/service/CommonManipulatorResultExtractorTest.java
@@ -10,7 +10,7 @@ import jakarta.inject.Inject;
 import org.jboss.pnc.api.reqour.dto.RemovedRepository;
 import org.jboss.pnc.api.reqour.dto.VersioningState;
 import org.jboss.pnc.reqour.adjust.model.ExecutionRootOverrides;
-import org.jboss.pnc.reqour.adjust.profiles.AdjustProfile;
+import org.jboss.pnc.reqour.common.profile.AdjustProfile;
 import org.junit.jupiter.api.Test;
 
 import java.net.URI;
